refactored cli, config and util.
