+++ package.json
@@ -18,7 +18,9 @@
     "resolve": "^1.1.6",
     "sane": "^1.2.0",
     "which": "^1.1.1",
-    "worker-farm": "^1.3.1"
+    "worker-farm": "^1.3.1",
+    "jest-constants": "^0.0.3",
+    "jest-util": "^0.0.2"
   },
   "devDependencies": {
     "babel-eslint": "^4.1.3",
