+++ packages/jest-constants/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "jest-constants",
+  "version": "0.0.3",
+  "main": "index.js"
+}
