+++ src/HasteModuleLoader/__tests__/HasteModuleLoader-jsdom-env-test.js
@@ -14,7 +14,7 @@ jest.autoMockOff();
 jasmine.DEFAULT_TIMEOUT_INTERVAL = 15000;
 
 const path = require('path');
-const utils = require('../../lib/utils');
+const normalizeConfig = require('../../config/normalize');
 
 describe('HasteModuleLoader', function() {
   let HasteModuleLoader;
@@ -22,7 +22,7 @@ describe('HasteModuleLoader', function() {
   let JSDOMEnvironment;
 
   const rootDir = path.join(__dirname, 'test_root');
-  const config = utils.normalizeConfig({
+  const config = normalizeConfig({
     cacheDirectory: global.CACHE_DIRECTORY,
     name: 'HasteModuleLoader-jsdom-env-tests',
     rootDir,
