+++ src/TestRunner.js
@@ -16,7 +16,7 @@ const getCacheKey = require('./lib/getCacheKey');
 const mkdirp = require('mkdirp');
 const path = require('path');
 const promisify = require('./lib/promisify');
-const utils = require('./lib/utils');
+const utils = require('jest-util');
 const workerFarm = require('worker-farm');
 
 const TEST_WORKER_PATH = require.resolve('./TestWorker');
