+++ src/config/defaults.js
@@ -1,0 +1,44 @@
+/**
+ * Copyright (c) 2014, Facebook, Inc. All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ */
+
+'use strict';
+
+const constant = require('jest-constants');
+const path = require('path');
+const utils = require('jest-util');
+
+module.exports = {
+  automock: true,
+  bail: false,
+  cacheDirectory: path.resolve(__dirname, '..', '..', '.haste_cache'),
+  coverageCollector: require.resolve('../IstanbulCollector'),
+  coverageReporters: ['json', 'text', 'lcov', 'clover'],
+  globals: {},
+  moduleFileExtensions: ['js', 'json', 'node'],
+  moduleLoader: require.resolve('../HasteModuleLoader/HasteModuleLoader'),
+  moduleResolver: require.resolve('../resolvers/HasteResolver'),
+  haste: {
+    providesModuleNodeModules: [],
+  },
+  modulePathIgnorePatterns: [],
+  moduleNameMapper: [],
+  testDirectoryName: '__tests__',
+  mocksPattern: '__mocks__',
+  testEnvironment: require.resolve('../environments/JSDOMEnvironment'),
+  testEnvData: {},
+  testFileExtensions: ['js'],
+  testPathDirs: ['<rootDir>'],
+  testPathIgnorePatterns: [utils.replacePathSepForRegex(constant.NODE_MODULES)],
+  testReporter: require.resolve('../reporters/IstanbulTestReporter'),
+  testURL: 'about:blank',
+  noHighlight: false,
+  noStackTrace: false,
+  preprocessCachingDisabled: false,
+  verbose: false,
+  useStderr: false,
+};
